move node-only types out of macros default entrypoint
They unnecessarily confuse typescript apps.
