+++ packages/compat/src/v1-addon.ts
@@ -16,7 +16,7 @@ import Options from './options';
 import walkSync from 'walk-sync';
 import ObserveTree from './observe-tree';
 import { Options as HTMLBarsOptions } from 'ember-cli-htmlbars';
-import { isEmbroiderMacrosPlugin } from '@embroider/macros';
+import { isEmbroiderMacrosPlugin } from '@embroider/macros/src/node';
 import { TransformOptions, PluginItem } from '@babel/core';
 import V1App from './v1-app';
 import modulesCompat from './modules-compat';
