+++ packages/core/src/app.ts
@@ -15,7 +15,7 @@ import { Asset, EmberAsset, ImplicitAssetPaths, InMemoryAsset, OnDiskAsset } fro
 import assertNever from 'assert-never';
 import SourceMapConcat from 'fast-sourcemap-concat';
 import Options from './options';
-import { MacrosConfig } from '@embroider/macros';
+import { MacrosConfig } from '@embroider/macros/src/node';
 import { PluginItem, TransformOptions } from '@babel/core';
 import { makePortable } from './portable-babel-config';
 import { TemplateCompilerPlugins } from '.';
