+++ packages/macros/src/ember-addon-main.ts
@@ -1,5 +1,5 @@
 import { join } from 'path';
-import { MacrosConfig, isEmbroiderMacrosPlugin } from '.';
+import { MacrosConfig, isEmbroiderMacrosPlugin } from './node';
 
 export = {
   name: '@embroider/macros',
