+++ packages/macros/src/index.ts
@@ -72,19 +72,3 @@ class Oops extends Error {
     this.params = params;
   }
 }
-
-// Entrypoint for managing the macro config within Node.
-export { default as MacrosConfig, Merger } from './macros-config';
-
-// Utility for detecting our babel and AST plugins.
-import { PluginItem } from '@babel/core';
-export function isEmbroiderMacrosPlugin(item: PluginItem) {
-  return (
-    (Array.isArray(item) &&
-      item.length > 1 &&
-      item[1] &&
-      typeof item[1] === 'object' &&
-      (item[1] as any).embroiderMacrosConfigMarker) ||
-    (item && typeof item === 'function' && (item as any).embroiderMacrosASTMarker)
-  );
-}
