+++ packages/macros/tests/babel/fail-build.test.ts
@@ -1,5 +1,5 @@
 import { allBabelVersions, runDefault } from './helpers';
-import { MacrosConfig } from '../..';
+import { MacrosConfig } from '../../src/node';
 
 describe(`fail build macro`, function () {
   allBabelVersions(function (transform: (code: string) => string, config: MacrosConfig) {
