+++ packages/macros/tests/glimmer/fail-build.test.ts
@@ -1,5 +1,5 @@
 import { templateTests } from './helpers';
-import { MacrosConfig } from '../..';
+import { MacrosConfig } from '../../src/node';
 
 describe(`macroFailBuild`, function () {
   templateTests(function (transform: (code: string) => string, config: MacrosConfig) {
